feature: allow to hide tag history to simplify the user interface
I would like to use docker-registry-ui to inform users about the available
tags for our images. The users of this interface are not interested in the
technical backgrounds how it works and I worry that it would confuse them.

Note: One needs to be aware, that this is data is available nonetheless. This
is only a user interface switch for hiding the button.