fix: no space between title and history icon in tag-history view
The ' ' space that is in the h2 in the template is swallowed up by the material icon. By adding a manual margin of the width of a normal space (1/4th character width).

Another solution would be to replace the space with a `&nbsp;`, but this adds the non-breaking space to the selection of text when highlighting words in the browser, which i did not prefer.