feature: improve contrast ratios in dark and light theme
This is done by inspecting the app using chromium-based dev tools. The dev-tools report the contrast ratios to their respective backgrounds. Every ratio – that i've found that was less then 4.5 – was updated using the automatic "next best" color that satisfies the ratio requirements for accessibility.

In some cases, i took the creative freedom to match some of the colors to match better in their hue based on hsl color theory. Please compare the screenshots in the associated PR. I tried to match the original style as well as possible.

I was only able to meet the contrast requirements by separating accent-text colors for content and header through the introduction of a new header-accent-text color.