feat: add `DOCKER_REGISTRY_UI_TITLE` and `ENABLE_VERSION_NOTIFICATION` (#369)
* feature: link title to main view

this way it is possible to go back to the main view, even if two layers deep, i.e. in the history view.

* feature: allow the customization of the title shown in the logo section

we already have REGISTRY_TITLE, but it is used as a reference to where the images come from: "Repositories of …". The UI however does not need to be used in a 1:1 context of the registry. Therefore, the project would benefit from this feature:

We therefore introduce "DOCKER_REGISTRY_UI_TITLE" as a new option, that optionally allows to change the title for more advanced customization and to enable more use cases for this software.

If it is not set, everything will stay the same as prior to this commit and the title will remain "Docker Registry UI"

Note, that the footer will always contain a reference to this project. I think it doesn't need to referenced twice, and makes room for some more customization.

* feature: allow to optionally disable the version check

Note: The default remains to have the version check enabled. It is though sometimes useful to disable it, in the context of customers or open source interests of your software not to bother with version updates to Docker Registry UI.

For example, if you use something like watchtower anyway, there is no need to bother with update notifications to anyone else other then the admin of the server.

* fix: wrongly named environment variable in error template

The environment variables named in the CATALOG_BRANCHING_CONFIGURATION error were wrong.

During that find, i found out that one needs to set both min and max to 0 to disable branching. I tried to make this clear in the README.md file.

* feat: update title and version notification options

Version check done every week instead of every days

---------

Co-authored-by: Joxit <joxit972@gmail.com>